revert main.py to original, combine two Policy to Policy2033766_2033525
